Rename _exposition to _legacy_exposition for clarity
